% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Reads in data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{location or name of the file to be read in}
}
\value{
This function returns the data formatted in data frame table
}
\description{
This function reads in data generated by the US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System.
}
\note{
If the data you are trying to pass doesn't exist in that location or is
incorrectly spell, function will return an error.
}
\examples{
fars_read(user/data/accident_2013.csv)
}
