% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads in data from multiple or even a single year in vector format.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list containing the different years}
}
\value{
This function returns the filename/s of the years passed. If a year is passed
that does not exist then NULL is returned.
}
\description{
This function reads in data from a vector containing a single or multiple years
and creates filenames for each year in the vector
}
\note{
An error will be thrown if a filename with a specific year does not exist.
}
\examples{
years <- c(2013, 2014, 2015)
fars_read_years(years)
}
