% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary of each year passed.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list containing the different years}
}
\value{
This function returns a summary of the accidents of each year passed in the parameter..
}
\description{
This function reads in a list of years and each year is
indivdually summarized of the number of fatal accidents.
}
\note{
An error will be thrown if a filename with a specific year does not exist
}
\examples{
\dontrun{
years <- c(2013, 2014)
fars_summarize_years(years)
}
}
