% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename based on year}
\usage{
make_filename(year)
}
\description{
This function takes in a year as its parameter and returns a character vector
with the name of the data set that is wanted in use.
}
\details{
@param year The year of the dataset

@return A character vector of the name of the dataset with associated year

@note parameter must be a valid year number or else when running function to find
 the file, file doesn't exist error will be given.

@examples
/dontrun{
 make_filename(2013)
 }
}
