% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot far a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer, the state you want to plot}

\item{year}{integer, the year you want to plot}
}
\value{
generate a map with a point for every location, return a NULL.
}
\description{
This function generate a map of the fars data releted to a specific year and state.
}
\examples{
\dontrun{
fars_map_state(12,2014)
}
}
