% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file into datatable}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is a character string with the file to be read}
}
\value{
a \code{data.frame} containing the csv data if file exists.
if file do not exists, the function do stop.
}
\description{
This function read the file and put it on a data.frame structure.
please note, this is a low-level function, check \code{\link{fars_read_years}} for a more userfriendly interface.
}
\examples{
\dontrun{
fars_read(make_filename(2017))
}
}
\seealso{
\code{\link{make_filename}}
}
