% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of years to summarize}
}
\value{
a tibble, summary with obs. for months in columsn and year in rows of fars data
}
\description{
Summarize fars data of many year
}
