% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate the filename for a particolar year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer, the year you want the file belong to}
}
\value{
a character string with filename for the requested year
}
\description{
Generate the filename for a particolar year
}
\examples{
make_filename(2016)
}
